+++ scripts/generic/action-grapple.js
@@ -1,20 +1,20 @@
-const version = "10.1";
+const version = "11.0";
 const optionName = "Grapple Action";
 const actorSizes = ["tiny","sm","med","lg", "huge", "grg"];
 
 try {
-	const defender = workflow.targets.first();
-	
+	const targetToken = workflow.targets.first();
+
 	// check the size diff
-	if (!isSizeEligible(token, defender)) {
+	if (!isSizeEligible(token, targetToken)) {
 		return ui.notifications.error(`${optionName} - target is too big to grapple`);
 	}
 
 	// Run the opposed skill check
-	const skilltoberolled = defender.actor.system.skills.ath.total < defender.actor.system.skills.acr.total ? "acr" : "ath";
+	const skilltoberolled = targetToken.actor.system.skills.ath.total < targetToken.actor.system.skills.acr.total ? "acr" : "ath";
 	let results = await game.MonksTokenBar.requestContestedRoll({token: token, request: 'skill:ath'},
-		{token: defender, request:`skill:${skilltoberolled}`},
-		{silent:true, fastForward:false, flavor: `${defender.name} tries to resist ${token.name}'s grapple attempt`});
+		{token: targetToken, request:`skill:${skilltoberolled}`},
+		{silent:true, fastForward:false, flavor: `${targetToken.name} tries to resist ${token.name}'s grapple attempt`});
 
 	let i=0;
 	while (results.flags['monks-tokenbar'][`token${token.id}`].passed === "waiting" && i < 30) {
@@ -24,10 +24,10 @@ try {
 
 	let result = results.flags["monks-tokenbar"][`token${token.id}`].passed;
 	if (result === "won" || result === "tied") {
-		await HomebrewMacros.applyGrappled(token, defender, 'opposed', null, null);
-		ChatMessage.create({'content': `${token.name} grapples ${defender.name}`})
+		await HomebrewMacros.applyGrappled(token, targetToken, 'opposed', null, null);
+		ChatMessage.create({'content': `${token.name} grapples ${targetToken.name}`})
 	}
-	
+
 } catch (err) {
     console.error(`${optionName}: ${version}`, err);
 }